Menu improvements
