+++ source/app/config/menu/edit.js
@@ -1,0 +1,37 @@
+"use strict";
+
+module.exports = [
+    {
+        label: "Undo",
+        accelerator: "CmdOrCtrl+Z",
+        role: "undo"
+    },
+    {
+        label: "Redo",
+        accelerator: "Shift+CmdOrCtrl+Z",
+        role: "redo"
+    },
+    {
+        type: "separator"
+    },
+    {
+        label: "Cut",
+        accelerator: "CmdOrCtrl+X",
+        role: "cut"
+    },
+    {
+        label: "Copy",
+        accelerator: "CmdOrCtrl+C",
+        role: "copy"
+    },
+    {
+        label: "Paste",
+        accelerator: "CmdOrCtrl+V",
+        role: "paste"
+    },
+    {
+        label: "Select All",
+        accelerator: "CmdOrCtrl+A",
+        role: "selectall"
+    }
+];
