+++ source/app/config/menu/help.js
@@ -1,0 +1,12 @@
+"use strict";
+
+const electron = require("electron");
+
+module.exports = [
+    {
+        label: "Learn More",
+        click: () => {
+            electron.shell.openExternal("http://buttercup.pw")
+        }
+    }
+];
