+++ source/app/config/menu/window.js
@@ -1,0 +1,21 @@
+"use strict";
+
+module.exports = [
+    {
+        label: "Minimize",
+        accelerator: "CmdOrCtrl+M",
+        role: "minimize"
+    },
+    {
+        label: "Close",
+        accelerator: "CmdOrCtrl+W",
+        role: "close"
+    },
+    {
+        type: "separator"
+    },
+    {
+        label: "Bring All to Front",
+        role: "front"
+    }
+];
