+++ source/app/lib/platform.js
@@ -1,0 +1,19 @@
+"use strict";
+
+class Platform {
+
+    static isOSX() {
+        return process.platform === 'darwin';
+    }
+
+    static isWindows() {
+        return process.platform === 'win32';
+    }
+
+    static isLinux() {
+        return process.platform === 'linux';
+    }
+
+}
+
+module.exports = Platform;
